Refactor Full Card Layout, Identify Items
